feat(pmu_d2255): add vmsd
